Added BMI calculations. - BMI is now automatically calculated through user entered height and weight. - Styling adjustments.

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -23,9 +23,9 @@ class CreateUsersTable extends Migration
             $table->string('password');
             $table->string('nickname')->nullable();
             $table->string('birthday');
-            $table->integer('height');
-            $table->integer('weight');
-            $table->integer('bmi');
+            $table->decimal('height');
+            $table->decimal('weight');
+            $table->decimal('bmi');
             $table->string('diet');
             $table->string('time_for_caffeine_intake')->nullable();
             $table->string('chronotype');
